issue-93.  * Fixed Auth component  * User store was separated from root
